+++ src/passwordler/decrypt_password.py
@@ -1,4 +1,5 @@
 import random
+from ._internals import original, getKeyMap
 
 def decrypt_password(encrypted_message, random_seed = 123):
     """
@@ -42,23 +43,16 @@ def decrypt_password(encrypted_message, random_seed = 123):
             'encrypted_message cannot be empty string')
     
     random.seed(random_seed)
-    
-    original = ['a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q',
-                'r','s','t','u','v','w','x','y','z','A','B','C','D','E','F','G','H',
-                'I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y',
-                'Z','é','è','à','ü','ö','ä','0','1','2','3','4','5','6','7','8','9',
-                '.',',',';',':','!','?','+','@','#','%','&','$','£','=','-','_',' ',]
 
-    encryption = original.copy()
-    random.shuffle(encryption)
-    
+    decryption = original.copy()
+    random.shuffle(decryption)
+
+    keyMap = getKeyMap(decryption, isDecryption=True)
     decrypted_msg = []
 
     for character in encrypted_message:
-        if character in original:
-            index = encryption.index(character)
-            new_char = original[index]
-            decrypted_msg.append(new_char)
+        if character in keyMap:
+            decrypted_msg.append(keyMap[character])
         else:
             decrypted_msg.append(character)
